+++ src/main/java/com/mraof/minestuck/block/SulfurBubbleBlock.java
@@ -2,12 +2,18 @@ package com.mraof.minestuck.block;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.BlockGetter;
-import net.minecraft.world.level.block.TallGrassBlock;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.BushBlock;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.Vec3;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
 
-public class SulfurBubbleBlock extends TallGrassBlock
+public class SulfurBubbleBlock extends BushBlock
 {
+	protected static final VoxelShape SHAPE = Block.box(3.0D, 0.0D, 3.0D, 13.0D, 13.0D, 13.0D);
+	
 	public SulfurBubbleBlock(BlockBehaviour.Properties properties)
 	{
 		super(properties);
@@ -17,5 +23,12 @@ public class SulfurBubbleBlock extends TallGrassBlock
 	protected boolean mayPlaceOn(BlockState state, BlockGetter level, BlockPos pos)
 	{
 		return state.is(MSBlocks.NATIVE_SULFUR.get());
+	}
+	
+	@Override
+	public VoxelShape getShape(BlockState pState, BlockGetter pLevel, BlockPos pPos, CollisionContext pContext)
+	{
+		Vec3 vec3 = pState.getOffset(pLevel, pPos);
+		return SHAPE.move(vec3.x, vec3.y, vec3.z);
 	}
 }