+++ src/main/java/com/mraof/minestuck/block/plant/BurnedFoliageBlock.java
@@ -1,16 +1,21 @@
 package com.mraof.minestuck.block.plant;
 
-import com.mraof.minestuck.block.MSBlocks;
+import com.mraof.minestuck.util.MSTags;
 import net.minecraft.core.BlockPos;
+import net.minecraft.tags.BlockTags;
 import net.minecraft.world.level.BlockGetter;
-import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.GrassBlock;
-import net.minecraft.world.level.block.TallGrassBlock;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.BushBlock;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.Vec3;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
 
-public class BurnedFoliageBlock extends TallGrassBlock
+public class BurnedFoliageBlock extends BushBlock
 {
+	protected static final VoxelShape SHAPE = Block.box(3.0D, 0.0D, 3.0D, 13.0D, 13.0D, 13.0D);
+	
 	public BurnedFoliageBlock(BlockBehaviour.Properties properties)
 	{
 		super(properties);
@@ -19,6 +24,13 @@ public class BurnedFoliageBlock extends TallGrassBlock
 	@Override
 	protected boolean mayPlaceOn(BlockState state, BlockGetter level, BlockPos pos)
 	{
-		return state.is(MSBlocks.IGNEOUS_STONE.get());
+		return state.is(MSTags.Blocks.HEAT_TERRAIN_BLOCKS) || state.is(BlockTags.DIRT);
+	}
+	
+	@Override
+	public VoxelShape getShape(BlockState pState, BlockGetter pLevel, BlockPos pPos, CollisionContext pContext)
+	{
+		Vec3 vec3 = pState.getOffset(pLevel, pPos);
+		return SHAPE.move(vec3.x, vec3.y, vec3.z);
 	}
 }