+++ src/main/java/com/mraof/minestuck/data/MSBlockStateProvider.java
@@ -146,6 +146,7 @@ public class MSBlockStateProvider extends BlockStateProvider
 		simpleBlockWithItem(MSBlocks.SHADE_STONE_CRUXITE_ORE);
 		simpleBlockWithItem(MSBlocks.PINK_STONE_CRUXITE_ORE);
 		simpleBlockWithItem(MSBlocks.MYCELIUM_STONE_CRUXITE_ORE);
+		simpleBlockWithItem(MSBlocks.UNCARVED_WOOD_CRUXITE_ORE);
 		
 		simpleBlockWithItem(MSBlocks.STONE_URANIUM_ORE);
 		simpleBlockWithItem(MSBlocks.DEEPSLATE_URANIUM_ORE);
@@ -165,6 +166,7 @@ public class MSBlockStateProvider extends BlockStateProvider
 		simpleBlockWithItem(MSBlocks.SHADE_STONE_URANIUM_ORE);
 		simpleBlockWithItem(MSBlocks.PINK_STONE_URANIUM_ORE);
 		simpleBlockWithItem(MSBlocks.MYCELIUM_STONE_URANIUM_ORE);
+		simpleBlockWithItem(MSBlocks.UNCARVED_WOOD_URANIUM_ORE);
 		
 		simpleBlockWithItem(MSBlocks.NETHERRACK_COAL_ORE);
 		simpleBlockWithItem(MSBlocks.SHADE_STONE_COAL_ORE);