+++ src/main/java/com/mraof/minestuck/item/MSCreativeTabs.java
@@ -632,6 +632,8 @@ public final class MSCreativeTabs
 		output.accept(MSItems.SHADE_STONE_CRUXITE_ORE.get());
 		output.accept(MSItems.PINK_STONE_CRUXITE_ORE.get());
 		output.accept(MSItems.MYCELIUM_STONE_CRUXITE_ORE.get());
+		output.accept(MSItems.UNCARVED_WOOD_CRUXITE_ORE.get());
+		
 		output.accept(MSItems.STONE_URANIUM_ORE.get());
 		output.accept(MSItems.DEEPSLATE_URANIUM_ORE.get());
 		output.accept(MSItems.NETHERRACK_URANIUM_ORE.get());
@@ -642,22 +644,31 @@ public final class MSCreativeTabs
 		output.accept(MSItems.SHADE_STONE_URANIUM_ORE.get());
 		output.accept(MSItems.PINK_STONE_URANIUM_ORE.get());
 		output.accept(MSItems.MYCELIUM_STONE_URANIUM_ORE.get());
+		output.accept(MSItems.UNCARVED_WOOD_URANIUM_ORE.get());
+		
 		output.accept(MSItems.NETHERRACK_COAL_ORE.get());
 		output.accept(MSItems.SHADE_STONE_COAL_ORE.get());
 		output.accept(MSItems.PINK_STONE_COAL_ORE.get());
+		
 		output.accept(MSItems.END_STONE_IRON_ORE.get());
 		output.accept(MSItems.SANDSTONE_IRON_ORE.get());
 		output.accept(MSItems.RED_SANDSTONE_IRON_ORE.get());
 		output.accept(MSItems.UNCARVED_WOOD_IRON_ORE.get());
+		
 		output.accept(MSItems.SANDSTONE_GOLD_ORE.get());
 		output.accept(MSItems.RED_SANDSTONE_GOLD_ORE.get());
 		output.accept(MSItems.SHADE_STONE_GOLD_ORE.get());
 		output.accept(MSItems.PINK_STONE_GOLD_ORE.get());
+		
 		output.accept(MSItems.END_STONE_REDSTONE_ORE.get());
 		output.accept(MSItems.UNCARVED_WOOD_REDSTONE_ORE.get());
+		
 		output.accept(MSItems.STONE_QUARTZ_ORE.get());
+		
 		output.accept(MSItems.PINK_STONE_LAPIS_ORE.get());
+		
 		output.accept(MSItems.PINK_STONE_DIAMOND_ORE.get());
+		
 		output.accept(MSItems.UNCARVED_WOOD_EMERALD_ORE.get());
 		
 		output.accept(MSItems.BLUE_DIRT.get());